requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
motor==3.6.0
pydantic==2.9.2
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
